fixup! fixup! fixup! fixup! Squash some integer truncation warnings

+++ xapian-core/tests/api_replicate.cc
@@ -110,11 +110,11 @@ truncated_copy(const string& srcpath,
 	FAIL_TEST("Open failed (when creating '" << destpath << "')");
     }
 
-    const int BUFSIZE = 1024;
+    const size_t BUFSIZE = 1024;
     char buf[BUFSIZE];
     size_t total_bytes = 0;
     while (tocopy > 0) {
-	size_t thiscopy = tocopy > BUFSIZE ? BUFSIZE : tocopy;
+	size_t thiscopy = tocopy > BUFSIZE ? BUFSIZE : size_t(tocopy);
 	size_t bytes = do_read(fdin, buf, thiscopy);
 	if (thiscopy != bytes) {
 	    FAIL_TEST("Couldn't read desired number of bytes from changeset");
